scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adbc-driver-postgresql>=1.4.0",
    "duckdb>=1.1.3",
    "ipykernel>=6.29.5",
    "kafka>=1.3.5",
    "polars>=1.20.0",
    "pyarrow>=19.0.0",
    "pyyaml>=6.0.2",
]
